Refactor for grpc user; add basic script for generating grpc code
