+++ .gitignore
@@ -23,3 +23,4 @@ data/*
 *.vim
 *.swp
 
+grpc-services/*
