--- models/api/api.go
@@ -1,17 +1,0 @@
-package api
-
-type Error struct {
-	Message string `json:"message,omitempty"`
-	Code    int    `json:"-"`
-}
-
-func (err Error) Error() string {
-	return err.Message
-}
-
-// SuccessResponse is the minimal response to send if the request went OK
-var SuccessResponse = struct {
-	Success bool `json:"success"`
-}{
-	Success: true,
-}
