--- models/api/shareRequest.go
@@ -1,5 +1,0 @@
-package api
-
-type ShareRequest struct {
-	UserID int `json:"userID"`
-}
