--- models/api/zipRequest.go
@@ -1,5 +1,0 @@
-package api
-
-type ZipRequest struct {
-	Paths []string `json:"paths"`
-}
