--- models/fileinfo.go
@@ -1,17 +1,0 @@
-package models
-
-import "time"
-
-type FileInfo struct {
-	ID             int       `storm:"id,increment"`
-	Path           string    `json:"path"`
-	Name           string    `json:"name"`
-	IsDir          bool      `json:"isDir"`
-	Size           int64     `json:"size"`
-	OwnerID        int       `json:"ownerID,omitempty"`
-	LastChanged    time.Time `json:"lastChanged,omitempty"`
-	MimeType       string    `json:"mimetype,omitempty"`
-	ParentID       int       `json:"parentID,omitempty"`
-	OriginalFileID int       `json:"originalFileID,omitempty"`
-	Starred        bool      `json:"starred"`
-}
