--- models/session.go
@@ -1,29 +1,0 @@
-package models
-
-import (
-	"fmt"
-	"strconv"
-	"time"
-)
-
-const SessionTokenLength = 32
-
-type Session struct {
-	UserID    int    `storm:"index"`
-	Token     string `storm:"id,unique"`
-	ExpiresAt time.Time
-}
-
-func (s Session) GetTokenString() string {
-	return fmt.Sprintf("%s%d", s.Token, s.UserID)
-}
-
-func ParseSessionTokenString(cookie string) (*Session, error) {
-	tok := cookie[:SessionTokenLength]
-	userIDStr := cookie[SessionTokenLength:]
-	userID, err := strconv.Atoi(userIDStr)
-	if err != nil {
-		return &Session{}, err
-	}
-	return &Session{UserID: userID, Token: tok}, nil
-}
